Added bond order assignment via ILP.
This patch adds functionality for automated bond type prediction according
to the ideas of the antechamber program. We changed the algorithm to an
exact and optimal version based on an ILP solver (we use lp_solve55 for
that task). The code should already work but still requires a number
of adaptations (in particular, lp_solve should be added to the contrib
package and to configure).

This code also contains contributions by Alexander Rurainski and Sabine Mueller.