support for multiple solutions in AssignBondOrderProcessor
We added a nested friend class ILPSolution storing all bond orders and the
resulting penaly score.
The ABOProcessor now tries to find a first solution. If the first solution
is a valid soultion, we look for additional solutions by prohibiting
each free bond to get assigned the bond order of the first solution.
Since the solver does not offer a unequality constraint type, we split the
problem into LE first_order-1 and GE first+1. The best solution is then
stored in the private variable solutions_.
Via a apply-method, the AtomContainer's bond orders can be set to any of
those solutions.
.... :-)