A-STAR implementation and reading the penalty values from an XML-datafile BondOrder.xml
we did some reorganisation in the class interface, e.g.
	its possible to get the penalty of a solution
	you can demand the next possible solution
and debugged the A-STAR-implementation.

the code was tested on 631 mmff94 training files, and in only 13 cases, the solution found by
antechamber was not in the set of optimal solutions found by the A-STAR.
We assume, that the algorithm of antechamber does not always find a minimal penalty

Still to do:
Tidy up and improve the GUI :-)