Further fixes for the GAFF atom typer.
It seems as if the thing now correctly assigns AMBER types. At least water
and AlaAla work as expected. Some stuff is still missing, though: mostly,
the recognition of AR5 (the other rings) and the handling of optional
numbers in atomic property strings. Also, all the stuff not occurring in
AMBER types but possibly in GAFF types needs to be tested.