Fixed bug #79 (double representations in edit mode).
This bug was caused by our attempt to create a ball and stick representation
for everything once we switch to edit mode. This was inconsistent and is
possibly very annoying, because we will never be able to consistently
remove those representations. Instead, we now merely switch to edit mode
without changing representations.

What we *should* probably do is open a dialog that warns the user if there
is no ball and stick representation for the currently highlighted container.