added a hybridisation processor
the processor assigns hybridisation states to all given atoms according the
choosen method.
The hybridisation processor has a method SMART_MATCHING, which tries to
match SMARTS as given in an xml-based inifile and assigns the corresponding
hybridisation states.
In the near future a GAFF_based method is planned to be added :-)