Added support for saving raytraceableGrid as .vrt raw file plus .vrt.vh text header. This is the format used directly by our external ray casters.
A note on that:
The data in RaytraceableGrid are stored as TRegularData3D<float>,
which defines "binaryWrite" to save its data to disk. However I needed
different "binaryWrite". So I added new method "binaryWriteRaw" to the general TRegularData3D<ValueType>
template class. However this method currently relies on the fact that
the data being written are floats. So I provided implementation only for
float specialization. So if you see error like "binaryWriteRaw has no body",
this means you are calling it from other than float specialization.