Initial version of a deprecation mechanism.
This commit introduces the macro BALL_DEPRECATED that can be used to mark
members of classes as deprecated. Usage:

    BALL_DEPRECATED void foo();

This version has only been tested under Linux and only with methods. It should work
on windows and with classes and variables, too. Problems might occur, though, please fix them
as they show up.