Add support for DESTDIR when installing
This introduces the DESTDIR variable which can be passed
to make install in order to install into a temporary
staging tree.
Additionally configure now tries to determine the library
directory (lib, lib64) used on the platform it is invoked on.
This behaviour can be enabled by passing --enable-libdir-detect
to configure.