Made recreation of lib[BALL,VIEW].objects unconditional.
While this might lead to unneccessary re-linking of libBALL and libVIEW
(e.g. when making two times in a row or just "make install"-ing), it
fixes a huge bug in the libtoolized version at least: the time stamp of
the .la - files did not change when object files changed, so the libs
were not updated.

This should also help when adding new files, since the libBALL.objects
and libVIEW.objects will be rebuilt anyhow.