Fix building the python modules with libtool
This uses the variable USE_LIBTOOL in order to distinguish
between a libtool and a non libtool build, so that the correct
object files are passed on to the linker. Currently this is kind
of a hack, but was the easiest solution.

For this the substitution of USE_LIBTOOL in configure.ac had to be
repaired