Fix neighbour box iterator for HashGrid3 (closes Bug #168).
This commit is a more or less complete rewrite of the BoxIterator
over neighbours. We used to try to keep track of non-empty neighbours
to speed up the loops, but this (a) did not work at all and (b) had
terrible runtime upon insert()/remove(). Now, each box has a
precomputed list of 27 box pointers, where the first one is always
guaranteed to be the box itself.