Split up aclocal into different m4 files
This allows to regenerate aclocal by issuing
aclocal -I m4 once in a while. For this to work out, the
libtool files had to be regenerated. The "bootstrap"
skript calles libtoolize, aclocal and autoconf in order
to prepare the buildsystem. The next step is to get rid
of all the automatically generated files...