Rework the plugin system.
This commit introduces the PluginHandler class, which takes care
of the plugin initialization. The PluginManager now contains
a list of registered PluginHandlers. These handlers are queried
via there method canHandle() if they are applicable to the passed
plugin type. If yes, then the Handler is tasked to start the plugin.
If multiple handlers can use this plugin, all of them are assigned to
start it.

In order to add a plugin to BALL it is now necessary to write the plugin
interface just as it used to be and additionally write at least one
PluginHandler, that can is capable of starting the plugin.