Added a ShortcutRegistry and made key sequences configurable.
This huge commit introduces a ShortcutRegistry and changes our previous
short cut definitions to a format that can be freely configured. We can currently
read and write shortcuts from/to text files that can be edited. In the next step, we
will add a GUI for editing short cuts on the fly.