Call main_display->makeCurrent() before exporting to POVRay
The switch to a buffered renderer infrastructure corrupted
the logic used by the POVRay renderer to obtain the camera
setup directly from OpenGL. As the render windows have an
orthographic setup, computing the camera setup resulted in
NaNs. This is fixed by switching the current context to the one
actually used for rendering the scene instead of painting
it to screen.