Split the python module
Split the python module into BALL and VIEW. This makes it possible
to build python bindings for libBALL without having to build libVIEW.

This commit mostly moves files around, adapts the buildsystem and
adds some cleanups where appropriate/needed.

An important change is the conditional building of serverWidget.sip,
which depends on the presence of the ASIO library.