Offscreen rendering works again, and has several new cool features!
We now use a tile rendering setup to be able to create nearly arbitrary large
exports. Also, we are no longer dependent on the availability of PixelBuffers,
even though these speed up things. If no PBuffer support is available, we render
the tiles in the main display, causing a brief flickering, but at least we can
generate the correct images.

The renderer also supports all the drawings/texts we use in various places, like
FPS strings and drawn 2D lines.

Offscreen rendering can also be added to the list of "regular" renderers, creating
an offline image at each frame!

Finally, we can now support rendering into arbitrary resolutions instead of using a
simple upscaling factor only.

TODO: - the changes need to be reflected in the StageSettings!
           - we need GUI support for hooking in a renderer that exports each frame to disk