Do not use yy_fatal_error in the parsers
yy_fatal_error by default calles exit() which is not something we want
in a shared library. Instead we now use our own implementation (called
ball_fatal_error) as an replacement. This implementation throws a
BALL::ParserException.

The function is written directly in the header file, as it will only be
included by .l/.y files and needs the type declarations which are only made
available through flex.