Make the NormalizeNamesProcessor smarter
Building on top of Marcel's fix for the naming standard problems,
the NormalizeNamesProcessor now tries to group Fragments by Chain or
NucleicAcid. This should yield a stabler estimate of the naming
standard. Orphan Fragments are still dealt with in a per Fragment
fashion.

Furthermore the monolithic finish() method has been split into smaller
chunks greatly enhancing maintainability and reusability of the code.

Note: I had to change some parts which suddenly ceased to work when
      splitting the processor into functions. Not entirely sure how
			those ever worked...