Work around a bug in SIP.
Types which are not used in a wrapped class, are not properly handled if
they were imported from another module. this lead to a severe bug in the
VIEW module.

To fix this a new dummy class for the list types used in extractors.sip
has been introduced, as well as a proper wrapper around std::list.