Revert "Added support for reading RNA molecules."
Unfortunately the FragmentDB _is_ used to obtain structures from.
Adding the hydroxy group of RNA would lead to regressions in all
code which is calling FragmentsBD::getFragment() with the name of
a nucleotide. Additionally the geometry of the hydroxy group was
not optimized such that the fragments cannot be used without further
postprocessing.

This reverts commit fdadd15dd6fdaad02728e3bd9f09e0ac99db9358.

Signed-off-by: Andreas Hildebrandt <anhi@bioinf.uni-sb.de>