Improved CPack-built Windows installer.
Apart from a few simple changes (logo, naming, ...), this commit
introduces our own copy of the NSIS - project template. This became
necessary to allow building installers with RequestExecutionLevel highest,
i.e. installers that make sense for admin and non-admin users.

Please note that as long as this is not included in CPack, we need to take
care to update the NSIS.template.in regularly.