Fix BISON parser generation on Windows.
As a hot candidate for the ugliest hack in BALL, I proudly present this
commit. Since Visual Studio 2010 external commands are now run in a single
batch file (apparently), we need to put a "call" before each invocation
of the magic runBison.bat - command. So we need to hack ourselves a way...