Add upload capabilities to SimpleDownloader
Yes, I know, this sounds stupid, however if you see it as being
complementary to requesting a resource via GET it becomes sensible
again.

For this change to work I introduced a class hierarchy for DLThread
in order to save code. We should reconsider the interface of this class
however as it is quite messy IMHO.