Fix potential crash of BALLView on startup and closing.
Loading the RotamerLibrary in the MolecularControl was recently
outsourced to a thread to improve startup times. Unfortunately, access
to the library was not synchronized, leading to race conditions and
regular crashes on startup and shutdown. This patch introduces a
ReadWriteLock for the RotamerLibrary and synchronizes access to it
accordingly.

Since the RotamerLibrary is now required in building the context menu
(to support changing rotamers), startup times have effectively increased
again. We could try to delay building of the context menu, or just add the
rotamers once they become available if this becomes a problem.