Added a new tool "PDBRMSDCalculator" for fast and efficient protein RMSD calculation
This tool computes the RMSD between proteins.

Parameters are either the second input file (i_query) whose type has to be specified (i_type) and can be either a single pdb, a dcd or a transformation file. Optional parameters are the rmsd type (-rmsd_type), and the type of atoms used for scoring a pose (-scope).

Output of this tool is a either the rmsd (in the pdb-pdb case) or a file (-o) containing the RMSD between the first pose and all other poses.