Stricter check for BALL_STD_STRING_HAS_CONST_ITERATOR_FUNCTIONS
In C++11 std::string::insert(it, 1, 'a') should return an iterator.
Unfortunately, this is not the case for MSVC 2010-2013 and GCC 4.8.2.

GCC also does not implement a version of insert that accepts a
const_iterator. MSVC however does so. Here, we run into the problem that
our wrapper expects that an iterator is returned.

In order to fix the chaos that ensues, we tighten the test to also check
for the appropriate return value.