Validate PDB insertion codes before accepting them
Currently Residue and Nucleotide accepted any character as PDB
insertion codes. This would wreak havoc once an invisible character
such as '\0' or '\n' were inserted. This patch ensures, that the
codes fall into the visible range and adds a convenience member
function for unsetting the current insertion code.

This fixes #544.