Modified RTTI function isKindOf to take a pointer argument.
On OSX 10.10 and later tests failed because a reference to
a deleted object was handed over to isKindOf. Taking the pointer
of that object led to a segmentation fault.
To prevent passing references to deleted objects, the reference
argument to be checked is replaced by the pointer of that object directly.