CMakeLists.txt: CMAKE_PREFIX_PATH overrides BALL_CONTRIB_PATH
With this change, CMAKE_PREFIX_PATH takes precedence before
BALL_CONTRIB_PATH. This is useful if a user wants to use a library
from contrib that is different from the one in the contrib package.
A prime example would be Qt, where a user could use a more recent
version instead of the one installed in contrib.