Fix XDRPersistenceManager_test
The test serialised a default constructed Bond, which was subsequently
deserialised. However, this means the Bond was serialised in an invalid
state which could trigger crashes upon deserialisation.

To fix the broken test just add two Atoms to the Bond.

Probably we should also make sure, that no objects in an invalid state
can be serialised...