Remove regularExpression.h include from file.h
Nothing in file.h needed a RegularExpression object. The dependency is
purely internal to the object file. However, it seems as if
regularExpression.h cannot be used with Google Test, as the Boost Regex
library and the POSIX regex library clash.

This is a nice example for: only include what you need.