Fix for AmberFF_test (issue #584)
This fix is only for the test itself.
Please read the discussion of issue 584 for more details on the actual problem
(selection changes followed by FF setup / updateEnergy).
In order to reduce the occurrence of test failures, according to Daniels suggestion,
I interspersed sleep time before calls to FF setup() or updateEnergy(). This reduces
the number of test failures roughly ten times.