BALLMacros.cmake: fix for dll finding in BALL_LIB_TO_DLL
If, as e.g. in case of libfftw-3.3.lib, the filename contained
more than one '.', not only the extension '.lib' was cut off by
get_filename_component(...) but also the last part of the filename itself
(in the example: '.3.lib') and the file has not been found. This has been fixed.