[BALL] Changed handling of adding compiler flags and definitions in the build system
Two CMake variables are introduced:

- BALL_PROJECT_COMPILE_FLAGS
- BALL_PROJECT_COMPILE_DEFNS

to store important compiler flags or definitions set by the BALL build system.
Here, important means: to be used also by external projects and thus to be stored
in BALLConfig.cmake. Unimportant flags are e.g. warnings that are not required.
Those are directly added to CMAKE_CXX_FLAGS or using ADD_DEFINITIONS() and they are
not stored in BALLConfig.cmake. Let's see if this works.

[BALL] Fixing compile flag/defns handling

[BALL] Fixing compile flag/defns handling

[BALL] Removed CMAKE_REQUIRED_DEFINITIONS settings

[BALL] Fixing CMAKE_REQUIRED_FLAGS handling