[Core] Quaternion class: replaced quaternion base class to Eigen
The boost base class for the BALL quaternion template class has been replaced
by the Eigen implementation for two reasons: keeping the interface, which is
not possible with boost versions >= 1.66 and getting rid of code that we have
to maintain that is provided by Eigen.

This commit also contains a fix for the function normalize(),
which used the squared norm instead of the euclidean norm.