[BALL] Fix Quaternion_test
When constructing axis-angles from quaternions, Eigen flips the rotation
axis if the real component of the quaternion is negative. Since this
does not affect the validity of the rotation, the result is still
correct. The test case has been adjusted to also accept the alternative
representation with flipped axis and modified angle of (2\pi -
old_angle).