[VIEW] Fix POVRenderer export
If the POVRenderer has been created with delete_ostrm_ = true (e.g.,
with a file name ctor), a call to clear() will reset the ostrm_ to
std::cout while keeping delete_ostrm_ = true.

Calling POVRenderer::clear() in init() caused the output stream to be
reset to std::cout, rendering it impossible to write output to a
different stream.