[VIEW] Move TilingRenderer to GLRenderer plugin (for now)
While the interface is general enough to be kept in VIEW, the actual
definitions only work with the old GLRenderer. Since this is currently
the only relevant consumer of the TilingRenderer, we will (for now) move
it to the GLRenderer plugin and reintroduce it in VIEW in a slightly
different version in the future.